# Speedup term-matching