--- a/man/glex_vi.Rd
+++ b/man/glex_vi.Rd
@@ -17,8 +17,6 @@ A \code{\link{data.table}} with columns:
 \item \code{degree} (\code{integer}): Degree of interaction of the \code{term}, with \code{1} being main effects,
 \code{2} being 2-degree interactions etc.
 \item \code{term} (\code{character}): Model term, e.g. main effect \code{x1} or interaction term \code{x1:x2}, \code{x1:x3:x5} etc.
-\item \code{term_list} (\code{list}): Same as \code{term} but as a list-column to enable filtering by specific variables without
-requiring to split by \code{:}.
 \item \code{class} (\code{factor}): For multiclass targets only: The associated target class. Lists all classes in the
 target, not limited to the majority vote.
 \item \code{m} (\code{numeric}): Average absolute contribution of \code{term}, see Details.