We sometimes need to identify interaction terms "belonging" to a main term (e.g. for SHAP calculation from interaction terms), but that can be cumbersome and slow for cases with many terms and high degree of interaction.
Previous implementation relied on a list-column of terms split at `:` and constructs `vapply` with `%in%` to find matchign terms.
With `find_term_matches` we can get rid of the superfluous list-column and gain some speed.
Combinatorics is still going to be painful for large `p` and `max_interaction` though.

``` r
p <- 4
degree <- 3

trm <- unlist(sapply(seq_len(degree), \(k) {
  combn(letters[1:p], k, simplify = FALSE) |>
    vapply(\(x) paste(x, collapse = ":"), character(1))
}))
trm
#>  [1] "a"     "b"     "c"     "d"     "a:b"   "a:c"   "a:d"   "b:c"   "b:d"  
#> [10] "c:d"   "a:b:c" "a:b:d" "a:c:d" "b:c:d"

glex:::find_term_matches("a", trm)
#> [1]  1  5  6  7 11 12 13

trm[glex:::find_term_matches("a", trm)]
#> [1] "a"     "a:b"   "a:c"   "a:d"   "a:b:c" "a:b:d" "a:c:d"
```

This also adds `get_degree` as a consistent helper to get the degree of interaction from a list of terms. Fixed substring splitting turns out to be faster than regex-ing the `:`:

``` r
# How many main terms
p <- 20

# generate interaction terms
terms <- unlist(sapply(seq_len(p), \(k) {
  combn(letters[1:p], k, simplify = FALSE) |>
    vapply(\(x) paste(x, collapse = ":"), character(1))
}))

head(terms, 3)
#> [1] "a" "b" "c"
tail(terms, 3)
#> [1] "a:c:d:e:f:g:h:i:j:k:l:m:n:o:p:q:r:s:t"  
#> [2] "b:c:d:e:f:g:h:i:j:k:l:m:n:o:p:q:r:s:t"  
#> [3] "a:b:c:d:e:f:g:h:i:j:k:l:m:n:o:p:q:r:s:t"

# as implemented in glex now
get_degree <- function(x) {
  lengths(strsplit(x, split = ":", fixed = TRUE))
}

# regex-based approach
get_degree_regex <- function(x) {
  lengths(regmatches(x, gregexpr(":", x))) + 1
}

bm <- bench::mark(
  get_degree(terms),
  get_degree_regex(terms)
)
#> Warning: Some expressions had a GC in every iteration; so filtering is
#> disabled.

bm
#> # A tibble: 2 × 6
#>   expression                   min   median `itr/sec` mem_alloc `gc/sec`
#>   <bch:expr>              <bch:tm> <bch:tm>     <dbl> <bch:byt>    <dbl>
#> 1 get_degree(terms)       277.18ms 277.84ms     3.60    12.24MB     0   
#> 2 get_degree_regex(terms)    9.05s    9.05s     0.110    8.15GB     3.54
plot(bm)
#> Loading required namespace: tidyr
```

![](https://i.imgur.com/8R5fiSB.png)<!-- -->